Update to follow WCAG consensus on rounding
As described in https://github.com/getflourish/Sketch-Color-Contrast-Analyser/issues/10 the values should not be rounded before comparing. Here I have moved the "rounding" to where it's being done just before showing it to the user. Instead of rounding, I floor the value to 1 decimal. Sadly JavaScript floating point math is tricky, so we have to do it using strings to avoid errors.